Add a streamlit UI
- pip install -r requirements.txt (into a virtual environment)
- streamlit run .\Waters2mzML-ui.py

The first time you run you'll need to set the location of your
msconvert.exe using "Find msconvert.exe" button. The location will be
saved in the .env file for subsequent runs.

The default raw_files & mzml_files folders can be changed and new values
will be saved in the .env file for subsequent runs.

Use the check boxes in the RAW Folder table to select the folders to
convert and press the "Process" button.

There is no input validation, yet!